feat(cp): refactor `cp` core util to use the new api design
