--- cmds/core/cp/cp_others.go
@@ -1,11 +1,0 @@
-// Copyright 2022 the u-root Authors. All rights reserved
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-//go:build !plan9
-
-package main
-
-import "syscall"
-
-const eNotDir = syscall.ENOTDIR
