--- cmds/core/cp/cp_plan9.go
@@ -1,9 +1,0 @@
-// Copyright 2022 the u-root Authors. All rights reserved
-// Use of this source code is governed by a BSD-style
-// license that can be found in the LICENSE file.
-
-package main
-
-import "fmt"
-
-var eNotDir = fmt.Errorf("not a directory")
