+++ cmds/exp/pox/pox.go
@@ -102,7 +102,7 @@ import (
 	"strings"
 	"syscall"
 
-	"github.com/u-root/u-root/pkg/cp"
+	"github.com/u-root/u-root/pkg/core/cp"
 	"github.com/u-root/u-root/pkg/ldd"
 	"github.com/u-root/u-root/pkg/mount"
 	"github.com/u-root/u-root/pkg/mount/loop"
