+++ pkg/core/cp/cp_test.go
@@ -16,9 +16,7 @@ import (
 	"golang.org/x/sys/unix"
 )
 
-var (
-	testdata = []byte("This is a test string")
-)
+var testdata = []byte("This is a test string")
 
 func TestCopySimple(t *testing.T) {
 	var err error
@@ -57,7 +55,7 @@ func TestCopySimple(t *testing.T) {
 			name:    "Success",
 			srcfile: srcfiles[0].Name(),
 			dstfile: dstfiles[0].Name(),
-			opt:     Default,
+			opt:     Options{},
 		},
 		{
 			name:    "SrcDstDirctoriesSuccess",
@@ -94,22 +92,11 @@ func TestCopySimple(t *testing.T) {
 		},
 	} {
 		t.Run(tt.name, func(t *testing.T) {
-			err := Copy(tt.srcfile, tt.dstfile)
-			if !errors.Is(err, tt.wantErr) {
-				t.Errorf("Test %q failed. Want: %q, Got: %q", tt.name, tt.wantErr, err)
-			}
-		})
-		//After every test with NoFollowSymlink we have to delete the created symlink
-		if strings.Contains(tt.dstfile, "symlink") {
-			os.Remove(tt.dstfile)
-		}
-
-		t.Run(tt.name, func(t *testing.T) {
 			if err := tt.opt.Copy(tt.srcfile, tt.dstfile); !errors.Is(err, tt.wantErr) {
 				t.Errorf("%q failed. Want: %q, Got: %q", tt.name, tt.wantErr, err)
 			}
 		})
-		//After every test with NoFollowSymlink we have to delete the created symlink
+		// After every test with NoFollowSymlink we have to delete the created symlink
 		if strings.Contains(tt.dstfile, "symlink") {
 			os.Remove(tt.dstfile)
 		}
@@ -119,17 +106,7 @@ func TestCopySimple(t *testing.T) {
 				t.Errorf("Test %q failed. Want: %q, Got: %q", tt.name, tt.wantErr, err)
 			}
 		})
-		//After every test with NoFollowSymlink we have to delete the created symlink
-		if strings.Contains(tt.dstfile, "symlink") {
-			os.Remove(tt.dstfile)
-		}
-
-		t.Run(tt.name, func(t *testing.T) {
-			if err := CopyTree(tt.srcfile, tt.dstfile); !errors.Is(err, tt.wantErr) {
-				t.Errorf("Test %q failed. Want: %q, Got: %q", tt.name, tt.wantErr, err)
-			}
-		})
-		//After every test with NoFollowSymlink we have to delete the created symlink
+		// After every test with NoFollowSymlink we have to delete the created symlink
 		if strings.Contains(tt.dstfile, "symlink") {
 			os.Remove(tt.dstfile)
 		}
