+++ pkg/securelaunch/launcher/vm_test.go
@@ -13,7 +13,7 @@ import (
 	"github.com/hugelgupf/vmtest/govmtest"
 	"github.com/hugelgupf/vmtest/guest"
 	"github.com/hugelgupf/vmtest/qemu"
-	"github.com/u-root/u-root/pkg/cp"
+	"github.com/u-root/u-root/pkg/core/cp"
 	slaunch "github.com/u-root/u-root/pkg/securelaunch"
 )
